Fix text width calculation and padding bugs
Default statuscolors patch includes the color escape characters in the
length computation, making it so any transition from one color to
another would add an extra space.

This removes the inclusion of that character in the width when drawing,
as well as removing it in the pre-computation of the width of the total
string.